refactor(card-table): split frame.py into toolbar mixin + placeholder helpers (#801)
Behavior-preserving decomposition of
widgets/panels/card_table_panel/frame.py (465 LOC) per the architecture
review in #801, following the package's frame/handlers/properties/protocol
mixin convention:

- toolbar.py: CardTablePanelToolbarMixin — view-mode buttons, pile-sort menu,
  and the printing dropdown + date dialog (#792). Owns the PRINTING_MODES /
  DATE_MODES imports so the printing concern lives in one place.
- placeholders.py: build_empty_state / build_loading_state free functions plus
  their _EMPTY_STATE_* / _ZONE_EMPTY_* constant tables. build_loading_state
  preserves the dynamic panel._label attribute consumed by handlers.show_loading.
- frame.py: construction core, public view-state API, and selection/zone-transfer
  bridge only. Mixin composed first with wx.Panel kept last in the MRO.
- protocol.py: extended with the toolbar mixin's self-surface (toolbar widgets,
  callbacks, _t, set_view_mode, set_pile_sort).

Code was moved, not rewritten; wx event-binding closures preserved intact.
